requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Face Detection
facenet-pytorch>=2.5.3

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Explainability
pytorch-grad-cam>=1.5.0

# Metrics & Scientific Computing
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
